requirements.txt
fastapi
uvicorn
paho-mqtt
influxdb-client
scikit-learn
numpy
